polish code with srpo
